refactor(report): extract Row types from interface.ts to data/rows.ts
Move 12 row-shape interfaces (TransactionSummaryDailyRow,
TransactionActiveUsersDailyRow, UserTransactionDailyRow,
UserTransactionAggregateRow, TransactionCommentRow,
UserProfileForReportRow, CommunityContextRow, DeepestChainRow,
DateRange, PeriodAggregateRow, RetentionAggregateRow,
CohortRetentionRow) from data/interface.ts into a sibling
data/rows.ts so the interface module can shrink to its repository
contract role. Callers (presenter / weeklyAggregator / repository /
service / weeklyAggregator.test.ts) import directly from rows.ts;
interface.ts re-exports the types so any external consumer keeps
resolving.

Mechanical move only — no shape changes to the interfaces themselves.
Sets up Phase B's repository split (the interface module gets
further partitioned next without dragging the row shapes through
each new file).
