+++ src/application/domain/report/data/repository.ts
@@ -7,12 +7,12 @@ import {
 } from "@prisma/client";
 import { injectable } from "tsyringe";
 import { IContext } from "@/types/server";
+import { IReportRepository } from "@/application/domain/report/data/interface";
 import {
   CohortRetentionRow,
   CommunityContextRow,
   DateRange,
   DeepestChainRow,
-  IReportRepository,
   PeriodAggregateRow,
   RetentionAggregateRow,
   TransactionActiveUsersDailyRow,
@@ -20,7 +20,7 @@ import {
   TransactionSummaryDailyRow,
   UserProfileForReportRow,
   UserTransactionAggregateRow,
-} from "@/application/domain/report/data/interface";
+} from "@/application/domain/report/data/rows";
 import {
   CommunitySummaryCursor,
   PrismaReport,
